% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{checkHSRevisionYearAndDigits}
\alias{checkHSRevisionYearAndDigits}
\title{Check HS Product Coding Request}
\usage{
checkHSRevisionYearAndDigits(hs_rev_year, hs_digits)
}
\arguments{
\item{hs_rev_year}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}

\item{hs_digits}{Number of digits in HS coding scheme (can be "4" or "6").}
}
\value{
TRUE if requested HS coding scheme is valid, FALSE otherwise.
}
\description{
\code{checkHSRevisionYearAndDigits} is a convenience function to make sure
that the user requests a valid coding scheme for OEC data.
}
