% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{checkHSRevisionYearAndDigitsForInputYear}
\alias{checkHSRevisionYearAndDigitsForInputYear}
\title{Check HS Product Coding Request For Input Year}
\usage{
checkHSRevisionYearAndDigitsForInputYear(hs_rev_year, hs_digits, input_year)
}
\arguments{
\item{hs_rev_year}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}

\item{hs_digits}{Number of digits in HS coding scheme (can be "4" or "6").}

\item{input_year}{Requested integer input year.}
}
\value{
TRUE if requested data is valid, FALSE otherwise.
}
\description{
\code{checkHSRevisionYearAndDigitsForInputYear} is a convenience function
to make sure that the user requests a valid coding scheme and input year for
OEC data.
}
