% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeComplexity}
\alias{computeComplexity}
\title{Compute Complexity Measure}
\usage{
computeComplexity(rca_panel)
}
\arguments{
\item{rca_panel}{Data frame output from \code{computeRCA}.}
}
\description{
\code{computeComplexity} takes country codes, product codes, and the RCA
matrix, and creates the panel that will be translated into the matrix M.
Each row represents a country and each column represents a product, in the
standard order. If a country has RCA > 1 for some product, the entry for
that country and product in M will be 1; otherwise it is 0.  For more
details about the matrix M, see Atlas p.25, Technical Box 2.2.
}
