% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeDiversity0}
\alias{computeDiversity0}
\title{Compute Diversity{0}}
\usage{
computeDiversity0(M_matrix)
}
\arguments{
\item{M_matrix}{Matrix output from \code{computeM}.}
}
\description{
\code{computeDiversity0} takes the complexity (M) matrix and computes the
initial values of diversity, called k_{c,0} in the Atlas. For more details,
see Atlas p.24, Technical Box 2.1.
}
