% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeM}
\alias{computeM}
\title{Compute Complexity ("M") Matrix}
\usage{
computeM(complexity_panel)
}
\arguments{
\item{complexity_panel}{Data frame output from \code{computeComplexity}.}
}
\description{
\code{computeM} takes the panel produced by computeComplexity and turns it
into a matrix with the correct row (country) and column (product) names
(codes). This matrix (designated in the Atlas using variable "M") is an
important input for the rest of the model.
}
