% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeMTilde}
\alias{computeMTilde}
\title{Compute M-Tilde}
\usage{
computeMTilde(M, diversity_0, ubiquity_0, flag)
}
\arguments{
\item{M}{Matrix output from \code{computeM}.}

\item{diversity_0}{Vector output from \code{computeDiversity0}.}

\item{ubiquity_0}{Vector output from \code{computeUbiquity0}.}

\item{flag}{Switch between product and country computations (can be "PCI" for
product complexity or "ECI" for economic, or country, complexity).}
}
\description{
\code{computeMTilde} takes the complexity (M) matrix, and the initial values
of diversity and ubiquity, and computes M-tilde_{pp'} as it would be called
in the Atlas. This is the precursor to computing the product and economic
complexity indices (PCIs/ECIs). For more details, see Atlas p.24, Technical
Box 2.1.
}
