% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeOutlook}
\alias{computeOutlook}
\title{Compute Outlook Measure}
\usage{
computeOutlook(M, pci, distance)
}
\arguments{
\item{M}{M matrix output of \code{\link{computeM}}.}

\item{pci}{PCI vector output of computeRanks(..., "pci").}

\item{distance}{Distance matrix output of \code{\link{computeDistance}}.}
}
\description{
\code{computeOutlook} takes complexity (M), the PCI, and distance and
computes a weighted measure of the total value of the products that a
country is not exporting, weighted by how close they are to what the country
is currently exporting. See Atlas p.54, Technical Box 5.4.
}
