% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeOutlookGain}
\alias{computeOutlookGain}
\title{Compute Outlook Gain Measure}
\usage{
computeOutlookGain(M, phi, pci, distance)
}
\arguments{
\item{M}{M matrix output of \code{\link{computeM}}.}

\item{phi}{Proximity matrix output of \code{\link{computeProximity}}.}

\item{pci}{PCI vector output of computeRanks(..., "pci").}

\item{distance}{Distance matrix output of \code{\link{computeDistance}}.}
}
\description{
\code{computeOutlookGain} takes complexity (M), proximity (phi), the PCI,
and distance, and computes the contribution of any one product in opening up
opportunities for producing more advanced products in any given country. See
Atlas p.54, Technical Box 5.4.
}
