% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeProximity}
\alias{computeProximity}
\title{Compute Proximity Measure}
\usage{
computeProximity(M, ubiquity_0)
}
\arguments{
\item{M}{Matrix output of \code{computeM}.}

\item{ubiquity_0}{Vector output of \code{computeUbiquity0}.}
}
\description{
\code{computeProximity} takes complexity (M) and the initial values of
ubiquity and computes a measure of the proximity (phi) between two products.
Proximity is defined as the conditional probability that a country exports
one product given that it already exports another. This measurement can be
used to enable visualizations of the product space. See Atlas p.52,
Technical Box 5.1.
}
