% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeRanks}
\alias{computeRanks}
\title{Compute Ranking Indicator (ECI/PCI)}
\usage{
computeRanks(M_tilde)
}
\arguments{
\item{M_tilde}{Matrix output of \code{computeMTilde}.}
}
\description{
\code{computeRanks} takes an M_tilde (either by country or by product)
and computes the appropriate ranking indicator (either ECI or PCI) by
finding the eigenvectors of M_tilde, taking the one associated with the
second eigenvalue, and normalizing its values. For more details, see Atlas
p.24, Technical Box 2.1.
}
