% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeRCA}
\alias{computeRCA}
\title{Compute Revealed Comparative Advantage}
\usage{
computeRCA(exports_panel)
}
\arguments{
\item{exports_panel}{Data frame containing export data.}
}
\description{
\code{computeRCA} takes the prepared export data and computes the revealed
comparative advantage, or RCA, matrix. Each row represents a country and each
column represents a product, in the standard order. For more details about
the RCA, see Atlas p.25, Technical Box 2.2.
}
