% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{computeUbiquity0}
\alias{computeUbiquity0}
\title{Compute Ubiquity{0}}
\usage{
computeUbiquity0(M_matrix)
}
\arguments{
\item{M_matrix}{Matrix output from \code{computeM}.}
}
\description{
\code{computeUbiquity0} takes the complexity (M) matrix and computes the
initial values of ubiquity, called k_{p,0} in the Atlas. For more details,
see Atlas p.24, Technical Box 2.1.
}
