% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{downloadAlbertaExports}
\alias{downloadAlbertaExports}
\title{Download Export Data From Alberta World Trade Atlas}
\usage{
downloadAlbertaExports()
}
\description{
\code{downloadAlbertaExports} downloads an appropriate exports dataset from
the Alberta World Trade Atlas.
}
