% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{downloadBACICountriesInfo}
\alias{downloadBACICountriesInfo}
\title{Download Country Coding Scheme From BACI}
\usage{
downloadBACICountriesInfo()
}
\description{
\code{downloadOECCountriesInfo} downloads the country coding scheme from the
BACI project.
}
