% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{downloadBACIExports}
\alias{downloadBACIExports}
\title{Download Export Data From OEC}
\usage{
downloadBACIExports(input_year)
}
\arguments{
\item{input_year}{integer year to download (can be 2012, 2013, or 2014)}
}
\description{
\code{downloadOECExports} downloads an appropriate exports dataset
from the Observatory of Economic Complexity.
}
