% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{downloadOECCountriesInfo}
\alias{downloadOECCountriesInfo}
\title{Download Country Coding Scheme From OEC}
\usage{
downloadOECCountriesInfo()
}
\description{
\code{downloadOECCountriesInfo} downloads the country coding scheme from the
Observatory of Economic Complexity.
}
