% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{generateBasicOutput}
\alias{generateBasicOutput}
\title{Generate Basic Output}
\usage{
generateBasicOutput(model)
}
\arguments{
\item{model}{Output from \code{\link{runModel}}.}
}
\description{
\code{generateBasicOutput} uses the model input to create some standardized
csv output. Output is dumped into the current working directory for now.
}
