% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{getValidInputYearsForHSRevisionYear}
\alias{getValidInputYearsForHSRevisionYear}
\title{Get Valid Input Years For HS Revision Year}
\usage{
getValidInputYearsForHSRevisionYear(hs_rev_year)
}
\arguments{
\item{hs_rev_year}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}
}
\description{
\code{getValidInputYearsForHSRevisionYear} is a convenience function
to get the appropriate input years existing in OEC data for a given HS
revision year.
}
