% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadAlbertaExportsPanel}
\alias{loadAlbertaExportsPanel}
\title{Load Export Data From Alberta World Trade Atlas}
\usage{
loadAlbertaExportsPanel()
}
\description{
\code{loadAlbertaExportsPanel} downloads an Alberta exports dataset from the
World Trade Atlas, if it isn't present in the /data folder, and returns the
result as a data frame.
}
