% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadBACICountriesInfoPanel}
\alias{loadBACICountriesInfoPanel}
\title{Load Country Coding Scheme From BACI}
\usage{
loadBACICountriesInfoPanel()
}
\description{
\code{loadBACICountriesInfoPanel} downloads the country coding scheme from the
BACI project, if it isn't present in the /data folder, and returns the
result as a data frame.
}
