% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadBACIExportsPanel}
\alias{loadBACIExportsPanel}
\title{Load Export Data From BACI}
\usage{
loadBACIExportsPanel()
}
\description{
\code{loadBACIExportsPanel} downloads a 2012-2014 exports dataset from the
BACI project, if it isn't present in the /data folder, and returns the
result as a data frame.
}
