% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadExportsPanel}
\alias{loadExportsPanel}
\title{Load Export Data}
\usage{
loadExportsPanel(name, hs_rev_year, hs_digits, input_year, ab_flag)
}
\arguments{
\item{name}{Name of data source (can be "oec" or "baci").}

\item{hs_rev_year}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}

\item{hs_digits}{Number of digits in HS coding scheme (can be "4" or "6").}

\item{input_year}{Year for which to run model (e.g. 2001, 2006).}

\item{ab_flag}{Set to TRUE to include data for Alberta, or FALSE to exclude.}
}
\description{
\code{loadExports Panel} downloads an appropriate exports dataset, if it
isn't present in the /data folder, and returns the result as a data frame.
}
