% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{loadModelObject}
\alias{loadModelObject}
\title{Load Model Object}
\usage{
loadModelObject(data_tag, product_code_rev, product_code_digit, input_year,
  ab_flag)
}
\arguments{
\item{data_tag}{Dataset to use (can be "baci" or "oec").}

\item{product_code_rev}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}

\item{product_code_digit}{Number of digits in HS coding scheme (can be "4" or "6").}

\item{input_year}{Year for which to run model (e.g. 2001, 2006).}

\item{ab_flag}{= TRUE (include AB data) or FALSE}
}
\description{
\code{loadModelObject} downloads the appropriate model from talentedco.in, if
it isn't present in the /data folder, and returns the result as a model
object.
}
