% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadOECCountriesInfoPanel}
\alias{loadOECCountriesInfoPanel}
\title{Load Country Coding Scheme From OEC}
\usage{
loadOECCountriesInfoPanel()
}
\description{
\code{loadOECCountriesInfoPanel} downloads the country coding scheme from the
Observatory of Economic Complexity, if it isn't present in the /data folder,
and returns the result as a data frame.
}
