% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadOECProductsInfoPanel}
\alias{loadOECProductsInfoPanel}
\title{Load Product Coding Scheme From OEC}
\usage{
loadOECProductsInfoPanel(hs_rev_year, hs_digits)
}
\arguments{
\item{hs_rev_year}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}

\item{hs_digits}{Number of digits in HS coding scheme (can be "4" or "6").}
}
\description{
\code{loadOECProductsInfoPanel} downloads the appropriate HS product coding
scheme from the Observatory of Economic Complexity, if it isn't present in
the /data folder, and returns the result as a data frame.
}
