% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{makeFocusedNetworkFromModel}
\alias{makeFocusedNetworkFromModel}
\title{Make Focused Network For Product Space From Economic Complexity Model}
\usage{
makeFocusedNetworkFromModel(model_object, focus_product_code, n_search_depth)
}
\arguments{
\item{model_object}{Model output of \code{\link{runModel}}.}

\item{focus_product_code}{String determining which product code to focus on.}

\item{n_search_depth}{Integer number of levels "away" from the focus product
to display.}
}
\description{
\code{makeFocusedNetworkFromModel} uses the model input, a target product
code, and a search depth limit  to create a version of the model focussed
on a single product, suitable for network visualization.
}
