% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{makeNetworkFromModel}
\alias{makeNetworkFromModel}
\title{Make Network For Product Space From Economic Complexity Model}
\usage{
makeNetworkFromModel(model_object, phi_cutoff_value)
}
\arguments{
\item{model_object}{Model output of \code{\link{runModel}}.}

\item{phi_cutoff_value}{Double cutoff value for proximity.}
}
\description{
\code{makeNetworkFromModel} uses the model input and a cutoff value of
proximity (phi) to create a version of the product space suitable for
network visualization.
}
