% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_obj}
\alias{model_obj}
\title{Placeholder dataset}
\format{An object of class \code{list} of length 5.}
\usage{
model_obj
}
\description{
A dataset containing minimal data, just hanging around here for now.
}
\keyword{datasets}
