% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prepareFocusedModelObject}
\alias{prepareFocusedModelObject}
\title{Make Focused Model Object For Product Space From Economic Complexity Model}
\usage{
prepareFocusedModelObject(model_object, focus_product_code, n_search_depth)
}
\arguments{
\item{model_object}{Model output of \code{\link{runModel}}.}

\item{focus_product_code}{String determining which product code to focus on.}

\item{n_search_depth}{Integer number of levels "away" from the focus product
to display.}
}
\description{
\code{prepareFocusedModelObject} uses the model input, a cutoff value of
proximity (phi), and a search depthj limit  to create a version of the
product space focussed on a single product, suitable for network
visualization.
}
