% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{runModel}
\alias{runModel}
\title{Run Standard OEC Model}
\usage{
runModel(data_tag, product_code_rev, product_code_digit, input_year, ab_flag)
}
\arguments{
\item{data_tag}{Dataset to use (can be "baci" or "oec").}

\item{product_code_rev}{HS coding scheme revision year (can be "1992", "1996",
"2002", or "2007").}

\item{product_code_digit}{Number of digits in HS coding scheme (can be "4" or "6").}

\item{input_year}{Year for which to run model (e.g. 2001, 2006).}

\item{ab_flag}{= TRUE (include AB data) or FALSE}
}
\description{
\code{runModel} runs the entire model program and returns all the results
wrapped in a single object. It also saves a copy of the model run to the
package data folder.
}
\details{
Note that for BACI data, the input_year is ignored when loading exports,
but applied when determining product names from codes.
}
