% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{runModelForAllOECData}
\alias{runModelForAllOECData}
\title{Run Model For All OEC Data}
\usage{
runModelForAllOECData()
}
\description{
\code{runModelForAllOECData} runs the model over all the Atlas data that
ecxplor knows about. (This takes a really long time!!!)
}
